api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
boto3==1.33.6
botocore==1.33.6
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.6
jmespath==1.0.1
joblib==1.3.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
patsy==0.5.4
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.8.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
statsmodels==0.14.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
